moved cluster into separate package
