+++ cluster/supercluster_topology.go
@@ -1,4 +1,4 @@
-package agent
+package cluster
 
 import (
 	"fmt"
